Move CostModel to own module
