+++ cardano-api/internal/Cardano/Api/ProtocolParameters.hs
@@ -1,12 +1,10 @@
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE RecordWildCards #-}
@@ -96,7 +94,7 @@ module Cardano.Api.ProtocolParameters (
   ) where
 
 import           Cardano.Api.Address
-import           Cardano.Api.Domain.Common
+import           Cardano.Api.Domain.CostModel
 import           Cardano.Api.Domain.PraosNonce
 import           Cardano.Api.Eras.Core
 import           Cardano.Api.Error
@@ -115,7 +113,6 @@ import           Cardano.Api.Utils
 import           Cardano.Api.Value
 
 import qualified Cardano.Binary as CBOR
-import qualified Cardano.Ledger.Alonzo.Language as Alonzo
 import qualified Cardano.Ledger.Alonzo.PParams as Ledger
 import qualified Cardano.Ledger.Alonzo.Scripts as Alonzo
 import qualified Cardano.Ledger.Api.Era as Ledger
@@ -130,8 +127,6 @@ import           Cardano.Slotting.Slot (EpochNo)
 import           Control.Monad
 import           Data.Aeson (FromJSON (..), ToJSON (..), object, withObject, (.!=), (.:), (.:?),
                    (.=))
-import           Data.Bifunctor (bimap, first)
-import           Data.Data (Data)
 import           Data.Either.Combinators (maybeToRight)
 import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
@@ -140,7 +135,6 @@ import           Data.Maybe.Strict (StrictMaybe (..))
 import           GHC.Generics
 import           Lens.Micro
 import           Numeric.Natural
-import           Text.PrettyBy.Default (display)
 
 
 -- -----------------------------------------------------------------------------
@@ -1068,62 +1062,6 @@ fromAlonzoPrices Alonzo.Prices{Alonzo.prSteps, Alonzo.prMem} =
   }
 
 
--- ----------------------------------------------------------------------------
--- Script cost models
---
-
-newtype CostModel = CostModel [Integer]
-  deriving (Eq, Show, Data)
-  deriving newtype (ToCBOR, FromCBOR)
-
-newtype CostModels = CostModels { unCostModels :: Map AnyPlutusScriptVersion CostModel }
-  deriving (Eq, Show)
-
-instance FromJSON CostModels where
-  parseJSON v = CostModels . fromAlonzoCostModels <$> parseJSON v
-
-instance ToJSON CostModels where
-  toJSON (CostModels costModels) =
-    case toAlonzoCostModels costModels of
-      Left err -> error $ displayError err
-      Right ledgerCostModels -> toJSON ledgerCostModels
-
-toAlonzoCostModels
-  :: Map AnyPlutusScriptVersion CostModel
-  -> Either ProtocolParametersConversionError Alonzo.CostModels
-toAlonzoCostModels m = do
-  f <- mapM conv $ Map.toList m
-  Right (Alonzo.emptyCostModels { Alonzo.costModelsValid = Map.fromList f })
- where
-  conv :: (AnyPlutusScriptVersion, CostModel) -> Either ProtocolParametersConversionError (Alonzo.Language, Alonzo.CostModel)
-  conv (anySVer, cModel) = do
-    alonzoCostModel <- toAlonzoCostModel cModel (toAlonzoScriptLanguage anySVer)
-    Right (toAlonzoScriptLanguage anySVer, alonzoCostModel)
-
-fromAlonzoCostModels
-  :: Alonzo.CostModels
-  -> Map AnyPlutusScriptVersion CostModel
-fromAlonzoCostModels (Alonzo.CostModels m _ _) =
-    Map.fromList
-  . map (bimap fromAlonzoScriptLanguage fromAlonzoCostModel)
-  $ Map.toList m
-
-toAlonzoScriptLanguage :: AnyPlutusScriptVersion -> Alonzo.Language
-toAlonzoScriptLanguage (AnyPlutusScriptVersion PlutusScriptV1) = Alonzo.PlutusV1
-toAlonzoScriptLanguage (AnyPlutusScriptVersion PlutusScriptV2) = Alonzo.PlutusV2
-toAlonzoScriptLanguage (AnyPlutusScriptVersion PlutusScriptV3) = Alonzo.PlutusV3
-
-fromAlonzoScriptLanguage :: Alonzo.Language -> AnyPlutusScriptVersion
-fromAlonzoScriptLanguage Alonzo.PlutusV1 = AnyPlutusScriptVersion PlutusScriptV1
-fromAlonzoScriptLanguage Alonzo.PlutusV2 = AnyPlutusScriptVersion PlutusScriptV2
-fromAlonzoScriptLanguage Alonzo.PlutusV3 = AnyPlutusScriptVersion PlutusScriptV3
-
-toAlonzoCostModel :: CostModel -> Alonzo.Language -> Either ProtocolParametersConversionError Alonzo.CostModel
-toAlonzoCostModel (CostModel m) l = first (PpceInvalidCostModel (CostModel m)) $ Alonzo.mkCostModel l m
-
-fromAlonzoCostModel :: Alonzo.CostModel -> CostModel
-fromAlonzoCostModel m = CostModel $ Alonzo.getCostModelParams m
-
 
 -- ----------------------------------------------------------------------------
 -- Proposals embedded in transactions to update protocol parameters
@@ -1906,18 +1844,3 @@ instance Error ProtocolParametersError where
     , "MaxBlockExUnits, MaxValueSize, CollateralPercent, MaxCollateralInputs. Did "
     , "you intend to use an Alonzo era protocol parameters value?"
     ]
-
-
-data ProtocolParametersConversionError
-  = PpceOutOfBounds !ProtocolParameterName !Rational
-  | PpceVersionInvalid !ProtocolParameterVersion
-  | PpceInvalidCostModel !CostModel !Alonzo.CostModelApplyError
-  | PpceMissingParameter !ProtocolParameterName
-  deriving (Eq, Show, Data)
-
-instance Error ProtocolParametersConversionError where
-  displayError = \case
-    PpceOutOfBounds name r -> "Value for '" <> name <> "' is outside of bounds: " <> show (fromRational r :: Double)
-    PpceVersionInvalid majorProtVer -> "Major protocol version is invalid: " <> show majorProtVer
-    PpceInvalidCostModel cm err -> "Invalid cost model: " <> display err <> " Cost model: " <> show cm
-    PpceMissingParameter name -> "Missing parameter: " <> name
